@@ -6,12 +6,12 @@ def siteUrl = 'https://gitlab.com/Trayis/SimpliRetro'      // Homepage URL of th
 def gitUrl = 'https://gitlab.com/Trayis/SimpliRetro.git'   // Git repository URL
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
 
     defaultConfig {
         minSdkVersion 17
-        targetSdkVersion 27
+        targetSdkVersion 28
         versionCode 1
         versionName "1.0"
     }
@@ -36,18 +36,18 @@ android {
 
 dependencies {
 
-    implementation "com.android.support:support-annotations:$androidLib"
+    implementation 'androidx.annotation:annotation:1.0.0'
 
     //Retrofit
     api "com.squareup.retrofit2:converter-gson:2.4.0"
     api "com.squareup.retrofit2:adapter-rxjava2:2.4.0"
-    api "io.reactivex.rxjava2:rxandroid:2.0.2"
+    api "io.reactivex.rxjava2:rxandroid:2.1.0"
 
     testImplementation 'junit:junit:4.12'
 }
 
 group = 'com.trayis'    // Maven Group ID for the artifact
-version = '0.2.1'       // This is the library version used when deploying the artifact
+version = '1.0.0'       // This is the library version used when deploying the artifact
 
 install {
     repositories.mavenInstaller {
